requirements.txt
azure-identity
requests
azure-storage-blob
